Add versions iOS and Darwin
Begin migration to new version "iOS" as alias for "IPhoneOS" and version
"Darwin" as an alias for "darwin".  This is first stage where OSX is
still defined for iOS (like the C TARGET_OS_ macros) and version OSX is
used for common MacOSX and iOS code.  Later goal is to convert all
common code to use version Darwin instead and eventually undefine OSX
for iOS targets.